readability edits to flow.cylc, no functional differences from the last commit, only style.
previously, nested Jinja2 conditionals were written like-

   {% if DO_SOMETHING %}
   {% if DO_THING_A %}
   thingA => something
   {% endif %}
   {% endif %}

This will now be written in the functionally equivalent manner with the Jinja2 blocks indented by a tab (4 spaces by my editors convention. The corresponding rose conf text will NOT be indented however, like so-

   {% if DO_SOMETHING %}
       {% if DO_THING_A %}
   thingA= something
       {% endif %}
   {% endif %}

In a similar manner the following else-if block written like-

   {% if DO_SOMETHING %}
   {% if DO_THING_A %}
   thingA => something
   {% elif DO_THING_B %}
   thingB => something
   {% elif DO_THING_C %}
   thingC => something
   {% endif %}
   {% endif %}

will now be written as-

   {% if DO_SOMETHING %}
       {% if DO_THING_A %}
   thingA => something
       {% elif DO_THING_B %}
   thingB => something
       {% elif DO_THING_C %}
   thingC => something
       {% endif %}
   {% endif %}

a for-loop block would be treated equivalently.