in trying to address needless CHUNK_B post-processing jobs, i noticed a problem in the logic of the following lines of code: https://gitlab.gfdl.noaa.gov/fre2/workflows/postprocessing/-/blob/224cafab260206e1865f0372c3f3ddab5eac6920/flow.cylc#L236-L265
notice that inside a Jinja2 block mandating `DO_NATIVE and DO_REGRID`, we further check for either `DO_NATIVE` and `DO_REGRID` individually
this does not make sense inside the nested if statement. so the surrounding requirement for both flags to be true has been removed.

Also notice that each if case in the else-if block ask for `CLEAN_WORK` to be true, so `CLEAN_WORK` is now the requirement to consider the entire block.

furthermore, `PP_CHUNK_A == PP_CHUNK_B` is an equivalent condition to `PP_CHUNK_B_COUNT==1` in the top if statement. not only this, but the coded
dependency graph chosen in the case of either `PP_CHUNK_B == PP_CHUNK_A` or `PP_CHUNK_A != HISTORY_SEGMENT` is equivalent, so these lines were
consolidated and rewritten to have equivalent functionality. Now the lines are put if the logical or of the above two evaluates to true,
 and the extra clean-shards line is added *only* if PP_CHUNK_COUNT == 1 is true after evaluating the or-statement.

committing these changes before proceeding to add in the ability to "turn off" the CHUNK_B processing.