new branch for tackling time-average generation in canopy
the python time-averaging script lives in fre-python-tools, located on github here:
https://github.com/NOAA-GFDL/fre-python-tools/tree/gen.time.avgs

development of *the time averaging script itself* + associated behavior will live on github.

this branch (2.frepythontools.gen.tim.avgs) is now about applying the fre-python-tools script
to the workflows defined in flow.cylc/rose-suite configurations, this will address
issue #2 (Generate timeaverages).

a new rose-suite variable has been added, `DO_TIMEAVGS`, and new task-family, `TIME-AVG`.

to start, i think i've identified a decent spot in the flow.cylc's dependency graph for
beginning to think about this. default behavior of time-averaging scripts in `nco`, `cdo`
and fre-nctools' `timavg.csh` all take in a history file with one variable's worth of
time-series, and averages it over the whole span, file-by-file. I.e., there's a one-to-one
relation between the input history file, and the output averaged history file.

this is a similar functional behavior to the `REFINE-DIAG` task-family, which produces a
history file with "refined diagnostics" from a given history file. in order to accomplish
this, it waits for `stage-history` tasks for finish, because those are the files it needs
then, `stage-history-refined` moves the output files elsewhere.

so, to start, we've defined a similar task-dependency for `TIME-AVG` to the `REFINE-DIAG`
family.