rename-split-to-pp routines kill the shards i need, and split-netcdf routines create them. so the proper dependency is probably like stage history all => regrid-xy<regrid> => split-netcdf-regrid<regrid => generate-time-averages<regrid> => rename-split-to-pp-regrid<regrid>
have some new commented out variables for local testing of generate-timeaverages script

reduced the number of components and turned off refinediag for now in flow.cylc/rose-suite.conf. this is just to speed up testing.

have to adjust the input directory via flow.cylc a little bit still. and the location of the python script. might be a better way to handle this.