trying to get somewhere with papiex op tags, and so defining a large workflow to take a look at generated workflow/task/scripts.
to make it nice an big so i can do a good job completely covering workflow definitions, i have DO_ANALYSIS=True.

upon rose macro --validate, I get an "assertion error", and that's it! nothing else.

traced the error to a few lines, which have a "if <some_object>" test in the form of "assert <some_object>". this is a good way to catch an error before wasting time running the workflow, but a bad way of informing the user of said error.

better to do a "if <some_object> is None: raise Exception(<blah>)" type thing.