cylc best practice asks for a `set -eou pipefail` within bash scripts, with one or two of those flags being on by default. this means anything with a non-zero exit status fails in the script- which is generally good, because usually that numeric status does indicate failure across many contexts as convention.
but cylc also knows that sometimes this is not desirable behavior and that the result of a program
can be complicated and nuanced, so i could do a "set +e" every dang time before a "export PAPIEX_TAGS=
...; <weird_command>" and after "unset PAPIEX_TAGS", but this is clunky looking, and will entail changing the logic structure of
the resultant workflow task scripts. not very amenable.

thankfully, i can chain together bash commands with '&&' in the if statement. so maybe i can tool the ops in the logical
statement itself.