this. worked.
for a given job_runner_handler in cylc, custom or otherwise, THE ONLY
text output allowed to be returned by the submit command is that of the
subprocess Popen call via cylc_subproc.

adding extra text output to the output, or the error, runs the risk of
breaking the regex search.

clean this up!!!!